[binary_storage] Fix clang-tidy errors in flash_partition and littlefs_mount
- Move esp_littlefs.h include from flash_partition.h to flash_partition.cpp
- Move lfs.h, LfsVfsContext and LfsVfsDir structs from littlefs_mount.h to
  littlefs_mount.cpp; use void* members with cast helpers in the class so the
  header has no dependency on managed component types
- Replace <errno.h> with <cerrno>
- Replace long with int32_t in fseek/ftell calls in seek()/tell()
- Add NOLINT for vfs_lfs_telldir/seekdir (long required by ESP-IDF VFS API)
- Fix bugprone-branch-clone in posix_flags_to_lfs (direct assignment)
- Fix bugprone-branch-clone in lseek whence handling (lookup array)
- Add NOLINT for malloc/free/strdup required by ESP-IDF VFS C API
- Replace index-based fd_paths cleanup loop with range-based for
- Fix test files: replace GPIOXX with platform-specific substitution variables

+++ esphome/components/binary_storage/flash_partition.cpp
@@ -3,6 +3,7 @@
 #ifdef USE_ESP_IDF
 
 #include "esphome/core/log.h"
+#include "esp_littlefs.h"
 #include <cstring>
 #include <cerrno>
 #include <sys/stat.h>
@@ -227,7 +228,7 @@ storage::StorageError FlashPartition::seek(storage::FileHandle *handle, size_t o
   if (handle == nullptr || !handle->in_use || handle->file == nullptr)
     return storage::StorageError::INVALID_ARGS;
 
-  if (fseek(handle->file, (long) offset, SEEK_SET) != 0)
+  if (fseek(handle->file, static_cast<int32_t>(offset), SEEK_SET) != 0)
     return storage::StorageError::INVALID_ARGS;
 
   return storage::StorageError::OK;
@@ -237,7 +238,7 @@ storage::StorageError FlashPartition::tell(storage::FileHandle *handle, size_t *
   if (handle == nullptr || !handle->in_use || handle->file == nullptr || position == nullptr)
     return storage::StorageError::INVALID_ARGS;
 
-  long pos = ftell(handle->file);
+  int32_t pos = ftell(handle->file);
   if (pos < 0)
     return storage::StorageError::READ_ERROR;
